+++ scripts/changelog.ts
@@ -82,7 +82,8 @@ async function exec(): Promise<void> {
     {
       added: [
         'Image plugin',
-        'Added `editable` prop to `Editor` for switching between edit and render mode'
+        'Added `editable` prop to `Editor` for switching between edit and render mode',
+        'Added optional `onPaste` callback to Plugins, which is called by the text plugin on paste event'
       ]
     }
   ])