+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Image plugin
 - Added `editable` prop to `Editor` for switching between edit and render mode
+- Added optional `onPaste` callback to Plugins, which is called by the text plugin on paste event
 
 ## [0.3.1](https://github.com/edtr-io/edtr-io/compare/0.3.0..0.3.1) - March 13, 2019